refactor(cli): Phase 0 hygiene — dup-source self-audit rule + cli.ts extraction (1.40.0)
kit 2.0 Phase 0 (auditable core). No behavior change.

self-audit R12-dup-source: flags cloud-sync conflict copies (the iCloud/Dropbox
' 2'/' 3' footgun: 'foo 2.ts', stray 'dist 2/' mirror dirs). Severity warn (local
junk, gitignored, never shipped) — turns a silent footgun into a surfaced advisory.
Pure isConflictCopyFile/isConflictCopyDir + catalog citation.

cli.ts god-file extraction begun: 7 command families (context, airgap, triage,
scan, verify-provenance, gha-audit, sbom) moved into src/commands/*; cli.ts 6482 ->
5662 lines, stays a dispatcher. Behavior-preserving; command-surface parity test
gates each move. Continues incrementally (ci/check/self-audit need shared CI-format
helpers lifted first; secrets/security/setup clusters next).

Verified under Node 22: build clean, command-surface parity passes, self-audit 0 fail.
(The one local `self-audit --fail-on-warning` test fails ONLY on this working tree
because R12 correctly warns on the untracked iCloud ' 2' junk present locally; a
clean CI checkout has no such files, so it passes there.)
