+++ src/cli.ts
@@ -111,15 +111,14 @@ import { promptConfirm } from "./utils/prompt.js";
 import { c } from "./utils/colors.js";
 import { gatherStatus } from "./status.js";
 import { KIT_FILE, resolveConfigPath } from "./cli-shared.js";
-import {
-  checkContext,
-  applyContext,
-  contextPrompt,
-  gatherLive,
-  suggestContextToml,
-  hasLockableContext,
-} from "./context-lock.js";
+import { gatherLive, suggestContextToml, hasLockableContext } from "./context-lock.js";
 import { cmdEnv } from "./commands/env.js";
+import { cmdContext } from "./commands/context.js";
+import { cmdAirgap } from "./commands/airgap.js";
+import { cmdScan } from "./commands/scan.js";
+import { cmdVerifyProvenance } from "./commands/verify-provenance.js";
+import { cmdGhaAudit } from "./commands/gha-audit.js";
+import { cmdSbom } from "./commands/sbom.js";
 import { cmdAuth } from "./commands/auth.js";
 import { cmdAudit } from "./commands/audit.js";
 import { cmdMcp } from "./commands/mcp.js";
@@ -135,8 +134,7 @@ import { cmdPlugin } from "./plugins-cli.js";
 import { cloneRepository } from "./clone.js";
 import { executeCommand } from "./run.js";
 import { listServices, openService } from "./open.js";
-import { gatherProjectContext } from "./context.js";
-import { runTriage, listTriageTools, type TriageType } from "./triage.js";
+import { cmdTriage } from "./commands/triage.js";
 import { parsePkgSpec, installPkg } from "./pkg.js";
 import { cmdMemory } from "./commands/memory.js";
 import { resolveKitRoot, runSelfAudit, SELF_AUDIT_RULES } from "./self-audit.js";
@@ -4590,184 +4588,6 @@ async function cmdOpen(): Promise<boolean> {
   }
 }
 
-async function cmdContextCheck(): Promise<boolean> {
-  const jsonMode = hasFlag(process.argv, "--json");
-  const config = await loadConfig(resolveConfigPath());
-  const findings = await checkContext(config.context);
-
-  if (jsonMode) {
-    console.log(JSON.stringify(findings, null, 2));
-    return findings.every((f) => f.status !== "mismatch");
-  }
-  if (!config.context) {
-    console.log(
-      `${c.dim}No [context] declared in .kit.toml — each CLI is unlocked from its account + project.${c.reset}`,
-    );
-    const suggestion = suggestContextToml(await gatherLive(process.cwd()));
-    if (suggestion) {
-      console.log(
-        `\nDetected here — add a ${c.bold}[context]${c.reset} block to .kit.toml to lock it:\n`,
-      );
-      console.log(suggestion);
-      console.log(
-        `\n${c.yellow}Verify each value is correct for THIS repo before trusting it.${c.reset} ${c.dim}kit detected the currently-active CLI state, which is exactly what the lock exists to question. Then re-run kit context check.${c.reset}`,
-      );
-    } else {
-      console.log(
-        `${c.dim}Add one to lock each CLI to its account + project (gcloud/vercel/github/git/npm).${c.reset}`,
-      );
-    }
-    return true;
-  }
-  console.log(`${c.bold}Context lock${c.reset}\n`);
-  for (const f of findings) {
-    const icon =
-      f.status === "ok"
-        ? `${c.green}✓${c.reset}`
-        : f.status === "mismatch"
-          ? `${c.red}✗${c.reset}`
-          : `${c.yellow}!${c.reset}`;
-    const actual = f.actual ?? `${c.dim}(unreadable)${c.reset}`;
-    const detail =
-      f.status === "ok"
-        ? `${c.dim}${f.expected}${c.reset}`
-        : `expected ${c.bold}${f.expected}${c.reset}, live ${c.bold}${actual}${c.reset}`;
-    console.log(`  ${icon} ${f.tool}.${f.field}  ${detail}`);
-  }
-  const mismatches = findings.filter((f) => f.status === "mismatch");
-  if (mismatches.length > 0) {
-    console.log(
-      `\n${c.red}${mismatches.length} mismatch(es) — you are NOT locked to the declared account/project. Do not run outward/destructive commands until this is green.${c.reset}`,
-    );
-  }
-  // Mismatch = non-zero exit so this can gate a hook / agent before acting.
-  // Unknown (tool absent / unreadable) does not block.
-  return mismatches.length === 0;
-}
-
-async function cmdContextUse(): Promise<boolean> {
-  const config = await loadConfig(resolveConfigPath());
-  if (!config.context) {
-    console.log(
-      `${c.dim}No [context] declared in .kit.toml. Add one to lock each CLI to its account + project.${c.reset}`,
-    );
-    return true;
-  }
-  console.log(`${c.bold}Context use${c.reset}\n`);
-  const results = await applyContext(config.context, process.cwd());
-  for (const r of results) {
-    const icon = r.ok ? `${c.green}✓${c.reset}` : `${c.red}✗${c.reset}`;
-    console.log(`  ${icon} ${r.step.tool}  ${c.dim}${r.step.describe}${c.reset}`);
-  }
-  // vercel/npm have no clean per-repo "active" state to switch — guide instead.
-  if (config.context.vercel) {
-    console.log(
-      `  ${c.dim}vercel: pass --scope per command or run \`vercel link\`; kit does not switch the dashboard link.${c.reset}`,
-    );
-  }
-  if (config.context.npm) {
-    console.log(
-      `  ${c.dim}npm: registry is global — set with \`npm config set registry <url>\` if it differs.${c.reset}`,
-    );
-  }
-  const failed = results.filter((r) => !r.ok).length;
-  if (failed > 0) {
-    console.log(
-      `\n${c.yellow}${failed} step(s) failed (tool not installed?). Run kit context check to verify.${c.reset}`,
-    );
-  }
-  return failed === 0;
-}
-
-async function cmdContext(): Promise<boolean> {
-  if (process.argv[3] === "check") return cmdContextCheck();
-  if (process.argv[3] === "use") return cmdContextUse();
-  // Fast, read-only PS1 indicator: `kit context --prompt` -> "[gcp:<project>]".
-  if (hasFlag(process.argv, "--prompt")) {
-    console.log(contextPrompt());
-    return true;
-  }
-
-  const jsonMode = hasFlag(process.argv, "--json");
-
-  try {
-    const config = await loadConfig(resolveConfigPath());
-    const context = await gatherProjectContext(config, process.cwd());
-
-    if (jsonMode) {
-      console.log(JSON.stringify(context, null, 2));
-    } else {
-      // Human-readable format
-      console.log(`${c.bold}Project Context${c.reset}\n`);
-
-      console.log(`${c.bold}Project: ${c.reset}${context.projectName}`);
-      console.log(`${c.bold}kit Version: ${c.reset}${context.kitVersion}`);
-      console.log(`${c.bold}Active Environment: ${c.reset}${context.activeEnvironment}`);
-      console.log();
-
-      if (context.detectedStack && Object.keys(context.detectedStack).length > 0) {
-        console.log(`${c.bold}Detected Stack:${c.reset}`);
-        for (const [key, value] of Object.entries(context.detectedStack)) {
-          console.log(`  ${key}: ${value}`);
-        }
-        console.log();
-      }
-
-      if (context.tools.length > 0) {
-        console.log(
-          `${c.bold}Tools (${context.tools.filter((t) => t.ok).length}/${context.tools.length}):${c.reset}`,
-        );
-        for (const tool of context.tools) {
-          const status = tool.ok ? `${c.green}✓${c.reset}` : `${c.red}✗${c.reset}`;
-          const version = tool.installed
-            ? ` ${c.dim}(${tool.installed})${c.reset}`
-            : " not installed";
-          console.log(`  ${status} ${tool.name}${version}`);
-        }
-        console.log();
-      }
-
-      if (context.services.length > 0) {
-        console.log(
-          `${c.bold}Services (${context.services.filter((s) => s.authenticated).length}/${context.services.length}):${c.reset}`,
-        );
-        for (const service of context.services) {
-          const status = service.authenticated ? `${c.green}✓${c.reset}` : `${c.red}✗${c.reset}`;
-          console.log(`  ${status} ${service.name}`);
-        }
-        console.log();
-      }
-
-      if (context.secrets.keys.length > 0) {
-        console.log(
-          `${c.bold}Secrets (${context.secrets.keys.filter((s) => s.available).length}/${context.secrets.keys.length}):${c.reset}`,
-        );
-        for (const secret of context.secrets.keys) {
-          const status = secret.available ? `${c.green}✓${c.reset}` : `${c.red}✗${c.reset}`;
-          console.log(`  ${status} ${secret.name}`);
-        }
-        console.log();
-      }
-
-      if (context.locks.length > 0) {
-        const allInSync = context.locks.every((l) => l.inSync);
-        console.log(
-          `${c.bold}Lock Files${allInSync ? ` ${c.green}(in sync)` : ` ${c.yellow}(drift detected)`}:${c.reset}`,
-        );
-        for (const lock of context.locks) {
-          const status = lock.inSync ? `${c.green}✓${c.reset}` : `${c.yellow}⚠${c.reset}`;
-          console.log(`  ${status} ${lock.category}: ${lock.detail}`);
-        }
-      }
-    }
-
-    return true;
-  } catch (err) {
-    console.error(`${c.red}Error: ${(err as Error).message}${c.reset}`);
-    return false;
-  }
-}
-
 async function cmdStatus(): Promise<boolean> {
   const items = await gatherStatus();
   if (hasFlag(process.argv, "--json")) {
@@ -5128,371 +4948,6 @@ async function cmdSentinelStatus(): Promise<boolean> {
   return true;
 }
 
-async function cmdVerifyProvenance(): Promise<boolean> {
-  const args = process.argv.slice(3);
-  // First positional that isn't a value-flag's argument = the artifact.
-  const VALUE_FLAGS = new Set(["--bundle", "--trusted-root", "--identity", "--issuer"]);
-  let artifact: string | undefined;
-  for (let i = 0; i < args.length; i++) {
-    if (VALUE_FLAGS.has(args[i])) {
-      i++;
-      continue;
-    }
-    if (!args[i].startsWith("-")) {
-      artifact = args[i];
-      break;
-    }
-  }
-  if (!artifact) {
-    console.error(
-      "usage: kit verify-provenance <artifact> --bundle <file> [--trusted-root <f>] [--identity <id>] [--issuer <iss>]",
-    );
-    return false;
-  }
-  const { resolveAirGap } = await import("./airgap/config.js");
-  const { verifyProvenance } = await import("./airgap/provenance.js");
-  const { resolveToolBin } = await import("./utils/resolveTool.js");
-  const { execFileNoThrow } = await import("./utils/execFileNoThrow.js");
-  // Config-free: verifying a provenance bundle needs the artifact + bundle, not
-  // .kit.toml. Missing config is not an error — only [air_gap] trust settings are
-  // read (optional). Mirrors `kit scan`.
-  const config = existsSync(resolveConfigPath())
-    ? await loadConfig(resolveConfigPath())
-    : ({} as kitConfig);
-  const ag = resolveAirGap(config.air_gap, process.env);
-  const res = await verifyProvenance(
-    {
-      artifact,
-      bundle: flagValue(args, "--bundle"),
-      trustedRoot: flagValue(args, "--trusted-root") ?? ag.provenanceTrustedRoot,
-      certIdentity: flagValue(args, "--identity") ?? ag.provenanceCertIdentity,
-      certIssuer: flagValue(args, "--issuer") ?? ag.provenanceCertIssuer,
-    },
-    {
-      resolveBin: (b) => resolveToolBin(b),
-      run: async (b, a) => {
-        const r = await execFileNoThrow(b, a, { timeout: 60_000 });
-        return { ok: r.ok, stdout: r.stdout, stderr: r.stderr };
-      },
-    },
-  );
-  if (res.ok) {
-    console.log(`${c.green}✓ provenance verified${c.reset}  ${c.dim}${artifact}${c.reset}`);
-    return true;
-  }
-  console.error(`${c.red}✗ ${res.reason}${c.reset}`);
-  return false;
-}
-
-async function cmdScan(): Promise<boolean> {
-  const jsonMode = hasFlag(process.argv, "--json");
-  const { runScanners, suppressBaselined, dedupKey, scanHealthGate, isLocalSemgrepConfig } =
-    await import("./scanners.js");
-  const { resolveToolBin } = await import("./utils/resolveTool.js");
-  const { execFileNoThrow } = await import("./utils/execFileNoThrow.js");
-  const { loadBaseline, baselineGet, baselineSet, saveBaseline } = await import("./baseline.js");
-  const { SCANNERS, airGapScanners } = await import("./scanners.js");
-  const cwd = process.cwd();
-  // Config-free by design: scanning is project-agnostic, so a missing .kit.toml is
-  // NOT an error — fall back to an empty config (air-gap + tooling tokens then come
-  // from env). This is what lets `kit scan` run in any repo without `kit init` first.
-  const configPath = resolveConfigPath();
-  const config = existsSync(configPath) ? await loadConfig(configPath) : ({} as kitConfig);
-
-  // Air-gap posture from `.kit.toml [air_gap]` + env (env overrides). When
-  // enabled, drop cloud-only scanners and run the rest against local DBs with no
-  // network. See docs/AIR_GAP.md.
-  const { resolveAirGap } = await import("./airgap/config.js");
-  const ag = resolveAirGap(config.air_gap, process.env);
-  const airgap = airGapScanners(SCANNERS, ag.enabled);
-  if (ag.enabled) {
-    console.log(
-      `${c.dim}air-gap mode: offline scanners only${airgap.dropped.length ? ` (skipping cloud-only: ${airgap.dropped.join(", ")})` : ""}${c.reset}`,
-    );
-    // Loud rejection: a registry KIT_SEMGREP_CONFIG (p/<pack>, r/<rule>, auto)
-    // would fetch rules from semgrep.dev = egress, so semgrep is NOT run
-    // air-gapped. Only a LOCAL ruleset path enables semgrep offline.
-    const semCfg = process.env.KIT_SEMGREP_CONFIG?.trim();
-    if (semCfg && !isLocalSemgrepConfig(semCfg)) {
-      console.error(
-        `${c.yellow}! air-gap: KIT_SEMGREP_CONFIG='${semCfg}' is a registry config (would fetch from the semgrep registry = egress) — semgrep is NOT run air-gapped. Point KIT_SEMGREP_CONFIG at a local ruleset path to enable it offline.${c.reset}`,
-      );
-    }
-    // Verified offline threat-data bundle (optional): point scanners at a
-    // signed, integrity-checked local DB set. Fail-closed — never scan against
-    // unverified threat data.
-    if (ag.threatDataDir) {
-      const tdDir = ag.threatDataDir;
-      const { verifyThreatData } = await import("./airgap/threat-data.js");
-      const pubRef = ag.threatDataPubkey ?? "";
-      const pubPem = pubRef && existsSync(pubRef) ? readFileSync(pubRef, "utf8") : pubRef;
-      if (!pubPem) {
-        console.error(
-          `${c.red}✗ air-gap threat-data dir set but no public key (threat_data_pubkey / KIT_THREAT_DATA_PUBKEY) — cannot verify (fail-closed)${c.reset}`,
-        );
-        return false;
-      }
-      const res = verifyThreatData({
-        dir: tdDir,
-        publicKeyPem: pubPem,
-        readFile: (rel) => readFileSync(resolve(tdDir, rel)),
-        resolvePath: (rel) => resolve(tdDir, rel),
-      });
-      if (!res.ok) {
-        console.error(
-          `${c.red}✗ threat-data bundle rejected: ${res.reason} (fail-closed)${c.reset}`,
-        );
-        return false;
-      }
-      Object.assign(airgap.env, res.env);
-      console.log(
-        `${c.dim}verified threat-data bundle: ${res.artifacts} artifact(s) signed + checksummed${c.reset}`,
-      );
-    }
-  }
-
-  // Resolve scanner tokens (SNYK_TOKEN, …) from a configured tooling Infisical
-  // project so `kit scan` works without an `infisical run` wrapper (#65). The
-  // value flows vault→subprocess env; it is never logged.
-  const toolingTokens: Record<string, string> = {};
-  const tooling = config.scan?.tooling;
-  if (tooling?.project_id) {
-    const { fetchInfisicalProjectSecrets } = await import("./secret-backends.js");
-    const map = await fetchInfisicalProjectSecrets({
-      project_id: tooling.project_id,
-      environment: tooling.env,
-    });
-    for (const s of SCANNERS) {
-      if (s.needsToken && !process.env[s.needsToken] && map.has(s.needsToken)) {
-        toolingTokens[s.needsToken] = map.get(s.needsToken)!;
-      }
-    }
-  }
-
-  const { merged, runs } = await runScanners(
-    {
-      resolve: (bin) => resolveToolBin(bin),
-      run: async (bin, args) => {
-        const r = await execFileNoThrow(bin, args, {
-          timeout: 180_000,
-          env: { ...process.env, ...airgap.env, ...toolingTokens },
-        });
-        return { ok: r.ok, stdout: r.stdout };
-      },
-      hasEnv: (name) => Boolean(process.env[name] || toolingTokens[name]),
-      detect: (markers) => markers.some((m) => existsSync(resolve(cwd, m))),
-    },
-    airgap.scanners,
-  );
-
-  // --update-baseline: freeze the current findings as accepted (#59 noise-reduction),
-  // reusing the shared .kit-baseline.json under a "scan" category.
-  if (hasFlag(process.argv, "--update-baseline")) {
-    const bl = await loadBaseline(cwd);
-    baselineSet(bl, "scan", "findings", merged.map(dedupKey));
-    await saveBaseline(bl, cwd);
-    console.log(
-      `${c.green}baseline updated${c.reset}  ${c.dim}${merged.length} scan finding(s) accepted into .kit-baseline.json${c.reset}`,
-    );
-    return true;
-  }
-
-  const accepted = new Set(baselineGet(await loadBaseline(cwd), "scan", "findings"));
-  const { kept: findings, suppressed } = suppressBaselined(merged, accepted);
-  const bad = findings.filter((m) => m.severity === "critical" || m.severity === "high").length;
-
-  // Scanner-HEALTH gate (separate from findings): a scanner that did not actually
-  // run (crashed/absent/no-token) must be able to force a non-zero exit so a green
-  // verdict never hides a broken scanner. Backward-compatible — a non-running
-  // scanner stays a loud WARN unless strict or a required scanner missed.
-  const strict = hasFlag(process.argv, "--strict") || envTruthy(process.env.KIT_CI_STRICT);
-  const requiredScanners = config.governance?.scan?.required_scanners ?? [];
-  const gate = scanHealthGate(runs, { strict, requiredScanners });
-  const overallOk = bad === 0 && gate.ok;
-
-  if (hasFlag(process.argv, "--sarif")) {
-    const { toSarif } = await import("./sbom.js");
-    console.log(JSON.stringify(toSarif(findings), null, 2));
-    return overallOk;
-  }
-  if (jsonMode) {
-    console.log(
-      JSON.stringify(
-        { runs, findings, suppressed, ok: overallOk, scannerGate: gate, strict },
-        null,
-        2,
-      ),
-    );
-    return overallOk;
-  }
-
-  const suppressNote = suppressed > 0 ? `  ${c.dim}(${suppressed} baselined)${c.reset}` : "";
-  console.log(
-    `${c.bold}kit scan${c.reset}  ${c.dim}external scanners → one merged verdict${c.reset}${suppressNote}`,
-  );
-  for (const r of runs) {
-    const mark =
-      r.status === "ran"
-        ? `${c.green}✓${c.reset}`
-        : r.status === "error"
-          ? `${c.red}✗${c.reset}`
-          : `${c.dim}−${c.reset}`;
-    const note = r.status === "ran" ? `${r.findings} finding(s)` : r.status;
-    console.log(`  ${mark} ${r.id}  ${c.dim}${note}${c.reset}`);
-  }
-  console.log("");
-  if (findings.length === 0) {
-    console.log(
-      `  ${c.green}no findings${suppressed > 0 ? " (after baseline)" : " from the scanners that ran"}${c.reset}`,
-    );
-  } else {
-    for (const m of findings) {
-      const sev = m.severity ?? "low";
-      const color =
-        sev === "critical" || sev === "high" ? c.red : sev === "medium" ? c.yellow : c.dim;
-      console.log(
-        `  ${color}${sev.toUpperCase().padEnd(8)}${c.reset} ${m.name}  ${c.dim}[${m.scanners.join("+")}]${c.reset}`,
-      );
-    }
-  }
-  if (bad > 0) console.log(`${c.red}${bad} high/critical${c.reset}`);
-  // Honest scanner-health reporting: name WHICH scanner did not run and why.
-  for (const w of gate.warnings) {
-    console.log(`  ${c.yellow}! ${w}${c.reset}`);
-  }
-  for (const fmsg of gate.failures) {
-    console.log(`  ${c.red}✗ ${fmsg}${c.reset}`);
-  }
-  if (gate.warnings.length > 0 && gate.ok) {
-    console.log(
-      `${c.dim}  (a non-running scanner is a warn; use --strict or [governance.scan] required_scanners to hard-fail)${c.reset}`,
-    );
-  }
-  return overallOk;
-}
-
-/**
- * `kit airgap verify` — prove the air-gap posture is egress-free: assert that
- * every scanner that WOULD run in air-gap mode resolves to a local artifact (no
- * cloud-only scanner, no registry/`auto` semgrep config). Read-only + deterministic.
- */
-async function cmdAirgap(): Promise<boolean> {
-  const sub = process.argv[3];
-  const jsonMode = hasFlag(process.argv, "--json");
-  if (sub !== "verify") {
-    console.error(`${c.red}usage: kit airgap verify${c.reset}`);
-    process.exitCode = 1;
-    return false;
-  }
-
-  const { SCANNERS, airGapScanners, verifyAirGapScanners, semgrepConfig } =
-    await import("./scanners.js");
-  const { resolveAirGap } = await import("./airgap/config.js");
-  const configPath = resolveConfigPath();
-  const config = existsSync(configPath) ? await loadConfig(configPath) : ({} as kitConfig);
-  const ag = resolveAirGap(config.air_gap, process.env);
-
-  // Evaluate the air-gap plan (what WOULD run with no egress), regardless of
-  // whether air-gap is currently enabled — the point is to PROVE the posture.
-  const plan = airGapScanners(SCANNERS, true, process.env);
-  const cfg = semgrepConfig(process.env);
-  const report = verifyAirGapScanners(plan.scanners, { semgrepConfig: cfg });
-
-  if (jsonMode) {
-    console.log(
-      JSON.stringify(
-        { ok: report.ok, enabled: ag.enabled, rows: report.rows, dropped: plan.dropped },
-        null,
-        2,
-      ),
-    );
-    return report.ok;
-  }
-
-  console.log(
-    `${c.bold}kit airgap verify${c.reset}  ${c.dim}every runnable scanner must be local (no egress)${c.reset}`,
-  );
-  if (!ag.enabled) {
-    console.log(
-      `${c.dim}  air-gap not enabled in .kit.toml / env — verifying what air-gap mode WOULD run${c.reset}`,
-    );
-  }
-  for (const row of report.rows) {
-    const mark = row.ok ? `${c.green}✓${c.reset}` : `${c.red}✗${c.reset}`;
-    console.log(`  ${mark} ${row.id}  ${c.dim}${row.detail}${c.reset}`);
-  }
-  for (const id of plan.dropped) {
-    console.log(
-      `  ${c.dim}− ${id}  dropped (cloud-only / registry — excluded from air-gap)${c.reset}`,
-    );
-  }
-  console.log("");
-  if (report.ok) {
-    console.log(
-      `  ${c.green}air-gap clean: all runnable scanners resolve to local artifacts${c.reset}`,
-    );
-  } else {
-    const bad = report.rows
-      .filter((r) => !r.ok)
-      .map((r) => r.id)
-      .join(", ");
-    console.log(`  ${c.red}air-gap NOT provable: ${bad} would egress${c.reset}`);
-  }
-  return report.ok;
-}
-
-async function cmdGhaAudit(): Promise<boolean> {
-  const jsonMode = hasFlag(process.argv, "--json");
-  const { runGhaAudit } = await import("./gha-audit.js");
-  const results = runGhaAudit(process.cwd());
-  const fails = results.filter((r) => r.status === "fail").length;
-
-  if (jsonMode) {
-    console.log(JSON.stringify({ ok: fails === 0, results }, null, 2));
-    return fails === 0;
-  }
-
-  console.log(`${c.bold}kit gha-audit${c.reset}  ${c.dim}.github/workflows hardening${c.reset}`);
-  for (const r of results) {
-    const mark =
-      r.status === "fail"
-        ? `${c.red}✗${c.reset}`
-        : r.status === "warn"
-          ? `${c.yellow}!${c.reset}`
-          : r.status === "skip"
-            ? `${c.dim}−${c.reset}`
-            : `${c.green}✓${c.reset}`;
-    console.log(`  ${mark} ${r.name}  ${c.dim}${r.detail}${c.reset}`);
-    if (r.suggestion) console.log(`      ${c.dim}${r.suggestion}${c.reset}`);
-  }
-  return fails === 0;
-}
-
-async function cmdSbom(): Promise<boolean> {
-  const fmt = (flagValue(process.argv, "--format") ?? "cyclonedx").toLowerCase();
-  if (fmt !== "cyclonedx" && fmt !== "spdx") {
-    console.error(`${c.red}usage: kit sbom [--format cyclonedx|spdx]${c.reset}`);
-    process.exitCode = 1;
-    return false;
-  }
-  const { lockComponents, toCycloneDX, toSpdx } = await import("./sbom.js");
-  const { parseLockPkgs } = await import("./supply-chain.js");
-  const cwd = process.cwd();
-  let lock: Parameters<typeof parseLockPkgs>[0];
-  try {
-    lock = JSON.parse(readFileSync(resolve(cwd, "package-lock.json"), "utf8"));
-  } catch {
-    console.error(`${c.red}no package-lock.json found — SBOM needs a committed lockfile${c.reset}`);
-    process.exitCode = 1;
-    return false;
-  }
-  const components = lockComponents(parseLockPkgs(lock));
-  console.log(
-    JSON.stringify(fmt === "spdx" ? toSpdx(components) : toCycloneDX(components), null, 2),
-  );
-  return true;
-}
-
 async function cmdWhoami(): Promise<boolean> {
   const jsonMode = hasFlag(process.argv, "--json");
 
@@ -5746,281 +5201,6 @@ async function cmdPkg(): Promise<boolean> {
   console.log(result.output);
 
   return result.installed;
-}
-
-async function cmdTriage(): Promise<boolean> {
-  const args = process.argv.slice(3);
-  const sandbox = hasFlag(args, "--sandbox");
-  const filtered = args.filter((a) => a !== "--sandbox");
-  const typeArg = filtered[0];
-  // `check-deps` is a pre-commit subcommand handled below before we narrow
-  // to TriageType; the rest of this function works against TriageType only.
-  if (typeArg === "check-deps") {
-    return cmdTriageCheckDeps();
-  }
-  const type = typeArg as TriageType;
-  const target = filtered[1];
-
-  if (!typeArg || typeArg === "--help" || typeArg === "-h") {
-    console.log(`${c.bold}kit triage${c.reset} — Security evaluation before installing packages\n`);
-    console.log("Usage:");
-    console.log("  kit triage docker <image>           Evaluate Docker image (CVE + sandbox)");
-    console.log("  kit triage npm <package>             Evaluate npm package (registry + GitHub)");
-    console.log(
-      "  kit triage npm <package> --sandbox   + offline tarball inspection (install-script + path-traversal scan, no code executed)",
-    );
-    console.log("  kit triage pip <package>             Evaluate PyPI package");
-    console.log(
-      "  kit triage brew <formula>            Evaluate Homebrew formula (resolves upstream repo, then repo-scores)",
-    );
-    console.log("  kit triage repo <github-url>         Evaluate GitHub repository");
-    console.log("  kit triage skill <path|name>         Evaluate Claude Code / agent skill");
-    console.log("  kit triage all <target>              Auto-detect and run all checks");
-    console.log("  kit triage tools                     Show installed security tools");
-    console.log(
-      "  kit triage check-deps                Pre-commit gate: fail if staged deps lack triage entries",
-    );
-    console.log("");
-    console.log("Examples:");
-    console.log("  kit triage npm express");
-    console.log("  kit triage npm node-ipc --sandbox");
-    console.log("  kit triage docker stalwartlabs/stalwart");
-    console.log("  kit triage skill searxng");
-    return true;
-  }
-
-  const validTypes = ["docker", "npm", "pip", "brew", "repo", "skill", "all", "tools"];
-  if (!validTypes.includes(type)) {
-    console.error(`${c.red}Unknown triage type: ${type}${c.reset}`);
-    console.error(`Valid types: ${validTypes.join(", ")}`);
-    return false;
-  }
-
-  if (type === "tools") {
-    const result = await listTriageTools();
-    console.log(result.output);
-    return true;
-  }
-
-  if (!target) {
-    console.error(`${c.red}Target required for triage ${type}${c.reset}`);
-    return false;
-  }
-
-  console.log(`${c.bold}Running triage on ${type}: ${target}${c.reset}\n`);
-  const result = await runTriage(type as TriageType, target);
-  console.log(result.output);
-
-  // Sandbox is currently only meaningful for npm packages — extends the
-  // registry-metadata check with offline tarball inspection.
-  let sandboxClean = true;
-  if (sandbox && type === "npm") {
-    const { triageNpmSandbox } = await import("./triage-sandbox.js");
-    console.log(`\n${c.bold}Sandbox inspection (offline):${c.reset}`);
-    try {
-      const sb = await triageNpmSandbox(target);
-      console.log(`  package: ${sb.package}${sb.version ? `@${sb.version}` : ""}`);
-      console.log(`  tarball: ${(sb.tarballSize / 1024).toFixed(1)} kB`);
-      console.log(`  install scripts: ${sb.hasInstallScripts ? "YES (review below)" : "none"}`);
-      if (sb.findings.length === 0) {
-        console.log(`  ${c.green}✓ no risk signals${c.reset}`);
-      } else {
-        const critical = sb.findings.filter((f) => f.severity === "critical");
-        if (critical.length > 0) sandboxClean = false;
-        for (const f of sb.findings) {
-          const tag =
-            f.severity === "critical"
-              ? `${c.red}[CRIT]${c.reset}`
-              : f.severity === "warn"
-                ? `${c.yellow}[WARN]${c.reset}`
-                : `[info]`;
-          console.log(`  ${tag} ${f.signal} — ${f.detail}`);
-        }
-      }
-    } catch (err) {
-      console.error(
-        `  ${c.red}sandbox failed: ${err instanceof Error ? err.message : String(err)}${c.reset}`,
-      );
-      sandboxClean = false;
-    }
-  } else if (sandbox && type !== "npm") {
-    console.log(
-      `\n${c.dim}(--sandbox currently only applies to npm; ignored for ${type})${c.reset}`,
-    );
-  }
-
-  const overall = result.passed && sandboxClean;
-  if (!overall) {
-    console.log(`\n${c.yellow}⚠️  Review warnings above before proceeding.${c.reset}`);
-  }
-
-  // Record the triage outcome so the pre-commit triage-check can verify
-  // that every newly added dep was evaluated. Failed runs aren't logged —
-  // a failed triage means the user explicitly chose not to install, so it
-  // shouldn't unblock the commit.
-  if (overall && target) {
-    await recordTriageRun(type, target, sandbox);
-  }
-
-  return overall;
-}
-
-const TRIAGE_LOG_FILE = ".kit-triage.jsonl";
-const TRIAGE_MAX_AGE_DAYS = 7;
-
-interface TriageLogEntry {
-  timestamp: string;
-  type: string;
-  target: string;
-  sandbox: boolean;
-  granter: string;
-}
-
-async function recordTriageRun(type: string, target: string, sandbox: boolean): Promise<void> {
-  const { appendFile } = await import("node:fs/promises");
-  const { resolve } = await import("node:path");
-  const entry: TriageLogEntry = {
-    timestamp: new Date().toISOString(),
-    type,
-    target,
-    sandbox,
-    granter: process.env.USER ?? "unknown",
-  };
-  try {
-    await appendFile(
-      resolve(process.cwd(), TRIAGE_LOG_FILE),
-      JSON.stringify(entry) + "\n",
-      "utf-8",
-    );
-  } catch (err) {
-    console.error(
-      `${c.dim}(triage-log append failed: ${err instanceof Error ? err.message : err})${c.reset}`,
-    );
-  }
-}
-
-/**
- * `kit triage check-deps` — pre-commit step. Diffs staged dep manifests
- * (package.json, requirements.txt, pyproject.toml) against the cached
- * versions in HEAD; any newly-added top-level package must have a triage
- * entry within the last `TRIAGE_MAX_AGE_DAYS` days. Exits non-zero on
- * missing entries so the commit fails.
- *
- * Intentionally does NOT walk transitive deps — that's the lockfile's
- * job. The pre-commit gate fires on what the developer is signing off on,
- * which is the manifest entries they wrote.
- */
-async function cmdTriageCheckDeps(): Promise<boolean> {
-  const { execFileSync } = await import("node:child_process");
-  const { readFile } = await import("node:fs/promises");
-  const { resolve } = await import("node:path");
-
-  const gitShow = (ref: string): string | null => {
-    try {
-      return execFileSync("git", ["show", ref], {
-        encoding: "utf-8",
-        stdio: ["pipe", "pipe", "ignore"],
-      });
-    } catch {
-      return null;
-    }
-  };
-
-  const newDeps: { ecosystem: "npm" | "pip"; name: string }[] = [];
-
-  // package.json — compare staged vs HEAD top-level dependencies.
-  const stagedPkg = gitShow(":package.json");
-  if (stagedPkg) {
-    try {
-      const headPkgRaw = gitShow("HEAD:package.json");
-      const staged = JSON.parse(stagedPkg) as {
-        dependencies?: Record<string, string>;
-        devDependencies?: Record<string, string>;
-      };
-      const head = headPkgRaw
-        ? (JSON.parse(headPkgRaw) as {
-            dependencies?: Record<string, string>;
-            devDependencies?: Record<string, string>;
-          })
-        : {};
-      const stagedDeps = { ...staged.dependencies, ...staged.devDependencies };
-      const headDeps = { ...head.dependencies, ...head.devDependencies };
-      for (const name of Object.keys(stagedDeps)) {
-        if (!(name in headDeps)) newDeps.push({ ecosystem: "npm", name });
-      }
-    } catch {
-      /* malformed package.json — skip */
-    }
-  }
-
-  // requirements.txt — line-oriented diff.
-  const stagedReq = gitShow(":requirements.txt");
-  if (stagedReq) {
-    const headReq = gitShow("HEAD:requirements.txt") ?? "";
-    const parseLines = (text: string) =>
-      new Set(
-        text
-          .split("\n")
-          .map((l) => l.trim())
-          .filter((l) => l && !l.startsWith("#"))
-          .map((l) => l.split(/[<>=!~]/)[0]!.trim()),
-      );
-    const stagedSet = parseLines(stagedReq);
-    const headSet = parseLines(headReq);
-    for (const name of stagedSet) {
-      if (!headSet.has(name)) newDeps.push({ ecosystem: "pip", name });
-    }
-  }
-
-  if (newDeps.length === 0) return true;
-
-  // Build a triage-log index of `<ecosystem>:<name>` → most-recent ISO ts.
-  const latest: Record<string, string> = {};
-  try {
-    const text = await readFile(resolve(process.cwd(), TRIAGE_LOG_FILE), "utf-8");
-    for (const line of text.split("\n")) {
-      if (!line.trim()) continue;
-      try {
-        const entry = JSON.parse(line) as TriageLogEntry;
-        const key = `${entry.type}:${entry.target}`;
-        const prev = latest[key];
-        if (!prev || entry.timestamp > prev) latest[key] = entry.timestamp;
-      } catch {
-        /* skip malformed line */
-      }
-    }
-  } catch {
-    /* no triage log yet */
-  }
-
-  const cutoff = Date.now() - TRIAGE_MAX_AGE_DAYS * 24 * 3600 * 1000;
-  const missing: string[] = [];
-  for (const dep of newDeps) {
-    const key = `${dep.ecosystem}:${dep.name}`;
-    const ts = latest[key];
-    // Parse first and reject non-finite: an unparseable timestamp yields NaN, and
-    // `NaN < cutoff` is false — treating a forged/garbage ts as "fresh triage"
-    // (fail-open). A missing or unparseable ts both count as "no valid triage".
-    const parsed = ts ? Date.parse(ts) : NaN;
-    if (!Number.isFinite(parsed) || parsed < cutoff) {
-      missing.push(
-        `  - ${dep.ecosystem}:${dep.name}  (run: kit triage ${dep.ecosystem} ${dep.name})`,
-      );
-    }
-  }
-
-  if (missing.length === 0) return true;
-
-  console.error(`${c.red}✗ Triage missing for ${missing.length} newly-added dep(s):${c.reset}`);
-  for (const line of missing) console.error(line);
-  console.error("");
-  console.error(
-    `${c.dim}Triage policy lives in CLAUDE.md. Run the suggested triage commands above,`,
-  );
-  console.error(
-    `then re-stage and commit. Bypass with --no-verify is recorded to ${SKIPPED_COMMITS_LOG}.${c.reset}`,
-  );
-  return false;
 }
 
 function cmdHelp(subcommand?: string): boolean {
